Migrates mitigation view to vuln view; deletes mitigation view
