--- View/UI/MitigationsNistMappingView.xaml.cs
@@ -1,15 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace Vulnerator.View.UI
-{
-    /// <summary>
-    /// Interaction logic for ContinuousMonitoringView.xaml
-    /// </summary>
-    public partial class MitigationsNistMappingView : UserControl
-    {
-        public MitigationsNistMappingView()
-        {
-            InitializeComponent();
-        }
-    }
-}
