+++ View/UI/ReportingView.xaml
@@ -1,12 +1,31 @@
 ﻿<UserControl x:Class="Vulnerator.View.UI.ReportingView"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:Controls="clr-namespace:MahApps.Metro.Controls;assembly=MahApps.Metro"
+             xmlns:Converters="clr-namespace:Vulnerator.View.Converter"
+             xmlns:IconPacks="http://metro.mahapps.com/winfx/xaml/iconpacks"
+             xmlns:LiveCharts="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
+             xmlns:ViewHelper="clr-namespace:Vulnerator.View.ViewHelper"
+             xmlns:ViewModel="clr-namespace:Vulnerator.ViewModel"
+             xmlns:WPFToolkit="clr-namespace:System.Windows.Controls;assembly=System.Windows.Controls.Layout.Toolkit"
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+             xmlns:dat="clr-namespace:System.Windows.Data;assembly=PresentationFramework"
+             xmlns:diag="clr-namespace:System.Diagnostics;assembly=WindowsBase"
+             xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
              xmlns:local="clr-namespace:Vulnerator.View.UI"
-             mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
-    <Grid>
-            
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+             xmlns:scm="clr-namespace:System.ComponentModel;assembly=WindowsBase"
+             HorizontalAlignment="Stretch"
+             VerticalAlignment="Stretch"
+             DataContext="{Binding Reporting, Source={StaticResource Locator}}"
+             mc:Ignorable="d">
+    <Grid Background="{StaticResource WindowBackgroundBrush}">
+        <Controls:MetroAnimatedSingleRowTabControl Margin="5"
+                                                   Controls:TabControlHelper.Underlined="SelectedTabItem">
+            <TabItem Header="Vulnerabilities"
+                     Style="{StaticResource CustomMetroTabItem}" />
+            <TabItem Header="Configuration Management"
+                     Style="{StaticResource CustomMetroTabItem}" />
+        </Controls:MetroAnimatedSingleRowTabControl>
     </Grid>
 </UserControl>
